+++ SVGDemo/SVGDemo/SVGDemo/Line.h
@@ -6,7 +6,7 @@ using namespace Gdiplus;
 
 class LINE : public Shape {
 private:
-	int x1, y1, x2, y2;
+	REAL x1, y1, x2, y2;
 public:
 	LINE(xml_node<>* node);
 	~LINE();
